project_requirements.txt
streamlit==1.28.1
scikit-learn==1.3.0
spacy==3.7.2
opencv-python==4.8.1.78
pandas==2.0.3
plotly==5.17.0
matplotlib==3.7.2
nltk==3.8.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
joblib==1.3.2
numpy==1.24.3
pillow==10.0.1